+++ packages/plugin-pha/src/runtime.ts
@@ -3,6 +3,7 @@ import type { StaticDataLoader } from '@ice/runtime/types';
 
 // 先写死timeout，未来如果有需要，可以通过dataLoader函数额外设置属性来支持
 const TIMEOUT = 3000;
+const APPWORKER = 'AppWorker';
 
 function timeout(interval: number) {
   return new Promise((_, reject) => {
@@ -19,9 +20,10 @@ function decorator(dataLoader: Function, index: Number = 0) {
     typeof window.pha.environment === 'object'
   ) {
     return (ctx: any) => {
+      // TODO： 这个链路需要打磨，主要涉及：是否要一个manifest一个worker？worker和webview通信协议设计（类似于小程序worker和webview的相互等待握手机制）
       const appWorkerPromsie = new Promise((resolve) => {
         const handler = (event) => {
-          if (event.origin === 'appWorker') {
+          if (event.origin === APPWORKER) {
             if (event.data && event.data.index === index) {
               resolve(event.data.data);
               // 注意只会使用一次